Test ETD download as a non logged in user
A test is added to download an unprotected etd as a non-logged in user.

In support of this:

The `:etd` factory is extended to accept a `pdf` keyword argument which accepts
a `Hyrax::UploadedFile`. If given the file will be attached to the work in a way
similar to what happens in the actor stack on ingest.

A `:perform_enqueued` spec metadata element is introduced to setup (and tear
down) the test queue adapter to perform jobs inline.