Test presense of title in ttl export
We test this at the presenter level, trusting the routing and controllers to do
their part.

The `EtdController` raises an unknown format error when `:ttl` is requested, so
it's not clear to me how conneg is being dispatched for .ttl responses. I think
we should eventually test that code locally as well if we can track down the
path.