Put identifier registration architecture in place
Adds a `DataciteActor` and a `DataciteRegisterJob` for asynchonous registration
of datacite ids.

Below these two, we have also added patterns ported from `epigaea` and its
Handle registration. We have an `IdentifierDispatcher` whose job is to assign registered
ids to a given object. This uses an (abstract) `IdentifierRegistrar` which
handles communication to the upstream identifier registration service.

An extremely simple dummy implementation of a `DataciteRegistrar` is also
included to support integration testing up to the `Actor` layer. This will need
to be extended later to support the Datacite API.