Reformat `department.yml` to correct terms list
The terms list contained duplicated items, each with *either* an `id` or a
`term`. This fixes each term to have both.

I believe it's appropriate *not* to test this change, since we want the
configuration to be correct. An alternative would be to provide a local
authority linter that could catch errors like this (e.g. `WARNING: some of your
authority terms do not have a required 'id'`.

Closes #149.