Pass XML metadata to DataCite when registering a DOI
Implement the initial POST requests for registering DOIs with DataCite.

For now we use a before filter to mark tests requiring the DataCite API as
pending. This is a temporary solution while we work on mocking the API
responses. If a valid looking datacite password is provided, we run the tests
against the real dev API instance.