Add feature tests for ETD visibility
Build out feature specs for visibility of public, institutional/authorized, and
private ETDs.

Hyrax's error messages for unauthorized access vary between explaining to the
user that they are 'not authorized' and that the page is 'private'. For now we
are satisfied with checking that the title doesn't appear on the page when
unauthorized access is attempted.

This commit omits specific tests for admin access to non-public visibilities. We
intend here to test the restrictions.