Complete registration through `DataciteRegistrar`
The registrar needs to send a URL to the connection's `#register` method to
complete the registration process.

`FakeDataciteConnection` is extended to support testing this, including by
adding spy attributes to check its iternal state. Unfortunately, due to Handle
system TTLs, there is no way to provide such immediate feedback in the real
implementation.