Move all `_attribute_rows` specs to `_metadata`
The idea behind this is that we don't actually care about the `_attribute_rows`
partial, and would be indifferent to see it squashed into `_metadata`. We can
test all the metadata for the show view in one place.

The downsides of this are: there isn't a spec file corresponding directly with
the `_attribute_rows.html.erb` partial, which we customize; hyrax's test suite
tests `_attribute_rows`, so this is a minor divergence from them.