Avoid retrying jobs on deserialization errors for Ldp::Gone
When a job tries to deserialize an object, it is possible for an error to
occur. One reason for these errors is that the object has been deleted and is no
longer accessible.

We know that for `ActiveFedora::Core` objects, this means the object is
permanently gone and that deserialization won't succeed in the future. We choose
to catch `DeserializationError` when the cause is `Ldp::Gone`, and log to avoid
further retries.