Document & refactor `ApplicationJob`
Make an initial pass at documenting the error handling and queue registration
strategies in `ApplicaitonJob`. We call out the intent to use `SideKiq` and
`Honeybadger` to manage error notifications and retries. Some details about how
to avoid retries are also provided, and the decision to avoid them on
`Ldp::Gone` is motivated.

The `queue_as` overwrite is fixed up slightly to handle the block setting
interface provided by `ActiveJob::Base`. We also switch to logging, rather than
raising errors in the production environment. This is never an error a user
should encounter.

Connected to #101.