Add MeSH autocomplete to subject field.
This commit changes the subject field to use MeSH terms for
autocompletion. The MeSH terms need to be installed on the
server using these instructions:

https://github.com/samvera/questioning_authority#medical-subject-headings-mesh

You will need to run `rake db:migate` after pulling in this commit.

This commit also sets up Capybara and adds a basic feature spec
for the field.

MeSH will require additional work so that Fedora stores the ID and
not the label. That is now issue #190.

Closes #124