Remove `:publisher` term from the Etd model
We had previously disabled `:publisher` on the form and views only. This removes
it at the model layer.

Unfortunately, `ActiveFedora` needs us to remove this from multiple `properties`
and `attributes` data structures, which all need to be syncrhonized. To do this
cleanly, we have added a `Schemas::GeneratedResourceSchemaStrategy#delete`
method. Ideally, this behavior would be pushed into `ActiveFedora` and some of
the synchronized data structures would be squashed out.