Add integration tests for importer task
An integration test for importing Etds via the rake task is added. Validation
for invalid CSV is also included.

Rake task tests are supported by an RSpec shared context which sets up the
task. The convention is that rake tasks have a top_level_description of the task
name, so a rubocop rule preventing this is disabled for the `spec/tasks`
directory.