Add a spec for creating a new version of a file
This partially tests file version creation by replacing the content of an
existing file with a new version. We test only that the contents are updated.

Some additional test setup may be necessary in order to get version creation to
work correctly in the test enviornment. Ideally we could test that versions are
trackable through the UI, and that old versions can be successfully restored.

Closes #213.