Add importer configuration for creator and file_path
`RecordImporter` needs to know which `User` to treat as the creator for imported
objects, and where on the file system to look for referenced file names. These
configuration options are added to `.env`, `config/importer.yml`, and
`Importer.config`. They are passed to `RecordImporter` by default on creation.

This work could be extended to allow overriding the configuration in the import
rake task by allowing that task to build a custom `RecordImporter`. For now, the
supporeted configuration is through `.env`.

Connected to #170.