Don't raise an error from Clamby when ClamAV virus is found
We are relying on the documented behavior that `#virus?` returns `true` when a
virus is found. Due to our configuration, we were instead encountering
`Exceptions::VirusDetected`; this happens even when a virus is not detected,
whenever the system call gives a non-0 exit code.

Changing to `error_file_virus: false` avoids the unexpected exceptions. We still
have a hair-trigger virus scanner that treats various system failure states as
positive virus scans. A ticket is open for this at
https://github.com/kobaltz/clamby/issues/12.